raspberry/requirements.txt
click==6.6
Django==1.9.7
django-bootstrap3==7.0.1
Flask==0.11.1
Flask-SocketIO==1.0b1
greenlet==0.4.10
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
Pillow==3.2.0
python-engineio==0.9.1
python-socketio==1.3
six==1.10.0
Werkzeug==0.11.10
